# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocumLink Kenya API — two-sided marketplace for medical staffing
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts/{shiftId}/applications:
    get:
      operationId: listShiftApplications
      tags: [applications]
      summary: List applications for a shift (clinic view)
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationList"
    post:
      operationId: applyToShift
      tags: [applications]
      summary: Apply to a shift (locum)
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyToShiftBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/shortlist:
    post:
      operationId: shortlistApplication
      tags: [applications]
      summary: Shortlist an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application shortlisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/confirm:
    post:
      operationId: confirmApplication
      tags: [applications]
      summary: Confirm an application (creates booking)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /applications/{id}/reject:
    post:
      operationId: rejectApplication
      tags: [applications]
      summary: Reject an application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /analytics/clinic:
    get:
      operationId: getClinicAnalytics
      tags: [analytics]
      summary: Get clinic HR analytics dashboard data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
      responses:
        "200":
          description: Clinic analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicAnalytics"
  /analytics/locum:
    get:
      operationId: getLocumAnalytics
      tags: [analytics]
      summary: Get locum earnings and performance analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
      responses:
        "200":
          description: Locum analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocumAnalytics"
...
  /admin/verification-queue:
    get:
      operationId: getVerificationQueue
      tags: [admin]
      summary: Get pending verification queue
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [locum, clinic]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Verification queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationQueue"
  /admin/locums/{id}/verify:
    post:
      operationId: adminVerifyLocum
      tags: [admin]
      summary: Approve or reject locum verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEntityBody"
      responses:
        "200":
          description: Locum verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/clinics/{id}/verify:
    post:
      operationId: adminVerifyClinic
      tags: [admin]
      summary: Approve or reject clinic verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEntityBody"
      responses:
        "200":
          description: Clinic verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/disputes/{id}/resolve:
    post:
      operationId: adminResolveDispute
      tags: [admin]
      summary: Resolve a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
...
    Specialty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        suggestedRateMin:
          type: integer
        suggestedRateMax:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - suggestedRateMin
        - suggestedRateMax
...
    Clinic:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        slug:
          type: string
        facilityType:
          type: string
        address:
          type: string
        subCounty:
          type: string
        county:
          type: string
        lat:
          type: string
        lng:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        verificationStatus:
          type: string
        payerScore:
          type: string
        totalShiftsPosted:
          type: integer
        totalShiftsFilled:
          type: integer
        logoUrl:
          type: string
        bio:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - facilityType
        - address
        - subCounty
        - verificationStatus
...
    Locum:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: string
        primarySpecialtyId:
          type: integer
        registrationNumber:
          type: string
        registrationBody:
          type: string
        yearsExperience:
          type: integer
        preferredRatePerShift:
          type: integer
        verificationStatus:
          type: string
        reliabilityScore:
          type: string
        totalShiftsCompleted:
          type: integer
        profilePhotoUrl:
          type: string
        isAvailableForUrgent:
          type: boolean
        subCounty:
          type: string
        county:
          type: string
        certExpiryDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        specialty:
          $ref: "#/components/schemas/Specialty"
      required:
        - id
        - firstName
        - lastName
        - registrationNumber
        - registrationBody
        - verificationStatus
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        specialtyId:
          type: integer
        title:
          type: string
        description:
          type: string
        shiftDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        rate:
          type: integer
        positionsAvailable:
          type: integer
        positionsFilled:
          type: integer
        status:
          type: string
        urgency:
          type: string
        minYearsExperience:
          type: integer
        specificRequirements:
          type: string
        insuranceCovered:
          type: string
        createdAt:
          type: string
          format: date-time
        clinic:
          $ref: "#/components/schemas/Clinic"
        specialty:
          $ref: "#/components/schemas/Specialty"
      required:
        - id
        - clinicId
        - specialtyId
        - title
        - shiftDate
        - startTime
        - endTime
        - rate
        - status
        - urgency
...
    Application:
      type: object
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        locumId:
          type: integer
        status:
          type: string
        coverMessage:
          type: string
        matchScore:
          type: string
        appliedAt:
          type: string
          format: date-time
        locum:
          $ref: "#/components/schemas/Locum"
        shift:
          $ref: "#/components/schemas/Shift"
      required:
        - id
        - shiftId
        - locumId
        - status
        - appliedAt
    ApplicationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
      required:
        - data
        - total
    ApplyToShiftBody:
      type: object
      properties:
        coverMessage:
          type: string

    # BOOKING SCHEMAS
    Booking:
      type: object
      properties:
        id:
          type: integer
        shiftId:
          type: integer
        locumId:
          type: integer
        applicationId:
          type: integer
        status:
          type: string
        contractUrl:
          type: string
        contractSignedByLocumAt:
          type: string
          format: date-time
        contractSignedByClinicAt:
          type: string
          format: date-time
        escrowAmount:
          type: integer
        checkedInAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - shiftId
        - locumId
        - status
...
    Dispute:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        disputeType:
          type: string
        description:
          type: string
        status:
          type: string
        resolutionNotes:
          type: string
        penaltyAppliedToLocum:
          type: integer
        penaltyAppliedToClinic:
          type: integer
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - disputeType
        - description
        - status
...
    ResolveDisputeBody:
      type: object
      properties:
        status:
          type: string
          enum: [resolved, escalated]
        resolutionNotes:
          type: string
        penaltyAppliedToLocum:
          type: integer
        penaltyAppliedToClinic:
          type: integer
      required:
        - status
        - resolutionNotes

    # NOTIFICATION SCHEMAS
...
    ClinicAnalytics:
      type: object
      properties:
        totalSpend:
          type: integer
        totalShiftsPosted:
          type: integer
        totalShiftsFilled:
          type: integer
        fillRate:
          type: number
        averageTimeToFill:
          type: number
        spendByMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              spend:
                type: integer
              shiftsPosted:
                type: integer
              shiftsFilled:
                type: integer
            required:
              - month
              - spend
              - shiftsPosted
              - shiftsFilled
        fillRateBySpecialty:
          type: array
          items:
            type: object
            properties:
              specialtyName:
                type: string
              posted:
                type: integer
              filled:
                type: integer
              fillRate:
                type: number
            required:
              - specialtyName
              - posted
              - filled
              - fillRate
        topLocums:
          type: array
          items:
            type: object
            properties:
              locumId:
                type: integer
              name:
                type: string
              shiftsCompleted:
                type: integer
              averageRating:
                type: number
            required:
              - locumId
              - name
              - shiftsCompleted
        hardestToFill:
          type: array
          items:
            type: object
            properties:
              specialtyName:
                type: string
              avgDaysToFill:
                type: number
            required:
              - specialtyName
              - avgDaysToFill
      required:
        - totalSpend
        - totalShiftsPosted
        - totalShiftsFilled
        - fillRate
    LocumAnalytics:
      type: object
      properties:
        totalEarnings:
          type: integer
        totalShiftsCompleted:
          type: integer
        averageRating:
          type: number
        reliabilityScore:
          type: string
        earningsByMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              earnings:
                type: integer
              shifts:
                type: integer
            required:
              - month
              - earnings
              - shifts
        recentActivity:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              type:
                type: string
              description:
                type: string
            required:
              - date
              - type
              - description
        upcomingShiftsCount:
          type: integer
        pendingApplicationsCount:
          type: integer
      required:
        - totalEarnings
        - totalShiftsCompleted
        - averageRating
        - reliabilityScore
...
    VerificationQueue:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              type:
                type: string
              name:
                type: string
              registrationNumber:
                type: string
              submittedAt:
                type: string
                format: date-time
              documents:
                type: array
                items:
                  type: string
            required:
              - id
              - type
              - name
              - submittedAt
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    VerifyEntityBody:
      type: object
      properties:
        status:
          type: string
          enum: [verified, rejected]
        notes:
          type: string
      required:
        - status